requirements.txt
fastapi
uvicorn
kafka-python
pyspark
snowflake-connector-python
